feat(Features/CustomTaxonomies): load custom taxonomies from json configuration files
Basically an adjusted copy of the CustomPostTypes feature. Since a lot of this overlaps with the
custom post type functionality, it can probably be refactored into one feature with two
sub-features. But since I am unsure that to keep using JSON files for these things even makes sense,
I didn't want to spend more time than necessary on it.

resolves #121